+++ internal/autodev/CHANGELOG.md
@@ -47,3 +47,15 @@
 
 ## [0.4.9+1781667780] - 2026-06-17
 - Test integration task
+
+## [0.4.9+1781673448] - 2026-06-17
+- Test integration task
+
+## [0.4.9+1781673685] - 2026-06-17
+- Test integration task
+
+## [0.4.9+1781674386] - 2026-06-17
+- Test integration task
+
+## [0.4.9+1781674543] - 2026-06-17
+- Test integration task